Add more ontology parsers
The order in which parsers are registered is important to ensure ontologies load correctly. For example, if OBOFormatOWLAPIParserFactory is registered before OWLFunctionalSyntaxOWLParserFactory, then ontologies written in functional syntax may fail to load because the OBO parser takes precedence. Adjusting the registration order ensures that the correct parser is applied first.